refactor and clean up test helpers
