+++ test/.rubocop.yml
@@ -5,3 +5,9 @@ Metrics/ClassLength:
 
 Metrics/MethodLength:
   Enabled: false
+
+Metrics/ModuleLength:
+  Enabled: false
+
+Metrics/AbcSize:
+  Enabled: false
