+++ test/integration/test_integration_helper.rb
@@ -1,14 +1,14 @@
 require 'test_helper'
 module TestIntegrationHelper
-  include TestHelper
   include PictureTag
+  include TestHelper
 
-  # PictureTag inherits from a basic Liquid tag, which helps it to interact with
-  # jeckyll but isn't necessary to test our code.
-
-  ContextStub = Struct.new(:environments, :registers)
-  SiteStub = Struct.new(:config, :data, :source, :dest)
-  TokenStub = Struct.new(:line_number, :locale)
+  def base_stubs
+    build_defaults
+    build_site_stub
+    build_context_stub
+    stub_liquid
+  end
 
   def tested(params = 'rms.jpg')
     Nokogiri::HTML(tested_base(params))
@@ -41,14 +41,6 @@ module TestIntegrationHelper
     '/tmp/jpt' + spx_url(width: width, format: format)
   end
 
-  def rms_file_array(widths, formats)
-    files = formats.collect do |f|
-      widths.collect { |w| rms_filename(width: w, format: f) }
-    end
-
-    files.flatten
-  end
-
   def std_spx_ss
     '/generated/spx-25-2e8bb3.jpg 25w,' \
        ' /generated/spx-50-2e8bb3.jpg 50w, /generated/spx-100-2e8bb3.jpg 100w'
@@ -59,184 +51,20 @@ module TestIntegrationHelper
       ' /generated/rms-50-46a48b.jpg 50w, /generated/rms-100-46a48b.jpg 100w'
   end
 
+  def rms_file_array(widths, formats)
+    files = formats.collect do |f|
+      widths.collect { |w| rms_filename(width: w, format: f) }
+    end
+
+    files.flatten
+  end
+
   def spx_file_array(widths, formats)
     files = formats.collect do |f|
       widths.collect { |w| spx_filename(width: w, format: f) }
     end
 
     files.flatten
-  end
-
-  def base_stubs
-    build_defaults
-    build_site_stub
-    build_context_stub
-    stub_liquid
-  end
-
-  # Since Nokogiri apparently still doesn't know about HTML5, we have to
-  # ignore 'invalid tag errors' which have an 801 code (I think):
-  def errors_ok?(output)
-    output.errors.none? { |e| e.code != 801 }
-  end
-
-  def nomarkdown_wrapped?(string)
-    # rubocop:disable Style/RegexpLiteral
-    start  = /^\{::nomarkdown\}/
-    finish = /\{:\/nomarkdown\}$/
-    # rubocop:enable Style/RegexpLiteral
-
-    string =~ start &&
-      string =~ finish &&
-      !string.include?("\n")
-  end
-
-  def build_context_stub
-    environments = [{ 'jekyll' => { 'environment' => @jekyll_env } }]
-    registers = {
-      site: @site,
-      page: @page
-    }
-
-    @context = ContextStub.new(environments, registers)
-  end
-
-  def build_defaults
-    @widths = [25, 50, 100]
-    @pconfig = {}
-    @pdata = {
-      'markup_presets' => {
-        'default' => {
-          'widths' => @widths
-        },
-
-        'auto' => {
-          'widths' => @widths,
-          'formats' => %w[webp original]
-        },
-
-        'data_auto' => {
-          'markup' => 'data_auto',
-          'widths' => @widths,
-          'formats' => %w[webp original]
-        },
-
-        'data_img' => {
-          'markup' => 'data_img',
-          'widths' => @widths
-        },
-
-        'data_picture' => {
-          'markup' => 'data_picture',
-          'widths' => @widths,
-          'formats' => %w[webp original]
-        },
-
-        'direct_url' => {
-          'markup' => 'direct_url',
-          'fallback_width' => 100
-        },
-
-        'img' => {
-          'markup' => 'img',
-          'widths' => @widths
-        },
-
-        'naked_srcset' => {
-          'markup' => 'naked_srcset',
-          'widths' => @widths
-        },
-
-        'sizes' => {
-          'sizes' => {
-            'mobile' => '80vw'
-          },
-          'size' => '50%'
-        },
-
-        'pixel_ratio' => {
-          'base_width' => 10,
-          'pixel_ratios' => [1, 2, 3]
-        },
-
-        'attributes' => {
-          'formats' => %w[webp original],
-          'widths' => @widths,
-          'attributes' => {
-            'parent' => 'class="parent"',
-            'alt' => 'Alternate Text',
-            'a' => 'class="anchor"',
-            'picture' => 'data-awesomeness="11"',
-            'source' => 'class="source"',
-            'img' => 'class="img"'
-          }
-        },
-
-        'link_source' => {
-          'widths' => @widths,
-          'link_source' => true
-        },
-
-        'media_widths' => {
-          'widths' => @widths,
-          'media_widths' => {
-            'mobile' => [10, 20, 30]
-          }
-        },
-
-        'data_noscript' => {
-          'markup' => 'data_img',
-          'widths' => @widths,
-          'noscript' => true
-        },
-
-        'fallback' => {
-          'widths' => @widths,
-          'fallback_width' => 35,
-          'fallback_format' => 'webp'
-        },
-
-        'nomarkdown' => {
-          'widths' => @widths,
-          'nomarkdown' => true
-        }
-      },
-
-      'media_presets' => {
-        'mobile' => 'max-width: 600px'
-      },
-
-      'too_large' => {
-        'widths' => [400, 600, 800],
-        'fallback_width' => 800
-      }
-
-    }
-
-    @jekyll_env = 'development'
-    @jconfig = { 'picture' => @pconfig,
-                 'keep_files' => [],
-                 'url' => 'example.com' }
-    @data = { 'picture' => @pdata }
-    @page = { 'ext' => 'html' }
-    @site_source = TEST_DIR
-    @site_dest = '/tmp/jpt'
-  end
-
-  def build_site_stub
-    @site = SiteStub.new(
-      @jconfig, @data, @site_source, @site_dest
-    )
-  end
-
-  def stub_liquid
-    Liquid::Template.stubs(:register_tag)
-
-    # Liquid::Template.parse(params).render(context)
-    template_stub = Object.new
-    Liquid::Template.stubs(:parse).with do |params|
-      template_stub.define_singleton_method(:render) { |_context| params }
-    end.returns(template_stub)
   end
 
   def cleanup_files
