+++ test/test_helper.rb
@@ -8,14 +8,15 @@ require 'mocha/minitest'
 require 'pry'
 require 'nokogiri'
 require 'jekyll_picture_tag'
+require_relative 'stubs/jekyll'
 
 module TestHelper
   TEST_DIR = __dir__
+  include JekyllStub
 
   ImageStruct = Struct.new(:width)
-
+  TokenStub = Struct.new(:line_number, :locale)
   ConfigStub = Struct.new(:source_dir)
-
   SourceImageStub = Struct.new(
     :base_name,
     :name,
@@ -39,4 +40,22 @@ module TestHelper
   SrcsetStub = Struct.new(
     :sizes, :to_s, :media, :mime_type, :media_attribute
   )
+
+  def nomarkdown_wrapped?(string)
+    # rubocop:disable Style/RegexpLiteral
+    start  = /^\{::nomarkdown\}/
+    finish = /\{:\/nomarkdown\}$/
+    # rubocop:enable Style/RegexpLiteral
+
+    string =~ start &&
+      string =~ finish &&
+      !string.include?("\n")
+  end
+
+  # Checks Nokogiri's html errors. Since it apparently still doesn't know about
+  # HTML5, we have to ignore 'invalid tag errors' which have an 801 code (I
+  # think):
+  def errors_ok?(output)
+    output.errors.none? { |e| e.code != 801 }
+  end
 end
